Finish remaining release readiness findings
