+++ IMPLEMENTATION-PLAN.md
@@ -25,9 +25,9 @@ The `release/v1.0.0` implementation wave delivers the first integrated stable ba
 - `M2` and `M4`: scan budgets, richer ignore profiles, data-driven detector modules, broad technology/vendor signature packs, generic repo artifact detection, repo-level demand fixtures, roadmap-archetype detector fixtures, policy coverage reporting, and benchmark coverage
 - `M3`: source utilization reporting, guarded docs/registry harvesters, VS Code Marketplace extension harvesting, and dependency-evidence npm/PyPI candidate generation
 - `M5` and `M6`: lifecycle terminology in wire plans, extension selection/install guidance, explicit VS Code extension install/verify/remove, shared MCP projection, asset prerequisite guidance, setup/doctor guidance, adapter-owned runtime preflight, host adapters, capability matrices, native project-local auto-wiring for Cursor, Zed, Claude Code, and Pi, and independent recommendation policies per registered host
-- `M7`: config, discovery demand profiling, source registry/index handling, source utilization reporting, catalog inspection, package/reference/local/GitHub/official-index harvesters, detector signatures, technology signatures, package candidates, wire, host adapters, path utilities, preflight/runtime validation, host-native implementation modules under `src/host-adapters/`, focused install-domain modules under `src/install/`, domain-specific manifest validators under `src/manifest-validation/`, and localized domain type modules under `src/types/`
+- `M7`: config, discovery demand profiling, source registry/index handling, source utilization reporting, catalog inspection, package/reference/local/GitHub/official-index harvesters, detector signatures, technology signatures, package candidates, wire, host adapters, path utilities, preflight/runtime validation, host-native implementation modules under `src/host-adapters/`, focused install-domain modules under `src/install/`, recommendation modules under `src/recommend/`, mirror modules under `src/mirror/`, domain-specific manifest validators under `src/manifest-validation/`, and localized domain type modules under `src/types/`
 
-Follow-up hardening after the merge addressed high-risk audit findings: scoped public package identity as `@ar27111994/agent-harness`, deterministic package allowlisting and packed-artifact smoke checks, package `main`/`exports` entry points, mutable state-root support through `--state-root` and `AGENT_HARNESS_STATE_ROOT`, safe mirror path resolution and file-manifest verification for multi-file artifacts, bounded mirror evidence file reads under approved local roots, guarded external response fetching with origin allowlists and byte limits, static public-provider allowlisting for optional AI enrichment, GitHub/repo content mirroring during mirror acquisition instead of wire time, validated PyPI metadata normalization, lazy VS Code settings path resolution after `.env` loading, resettable process-local GitHub state, explicit recommendation execution in the workspace and full rebuild pipelines, registry-driven recommendation-host enumeration, adapter-derived install bundle lock discovery, safe preview-by-default wire mode, non-destructive VS Code/OpenCode managed sections, quarantine review commands with audit logging, optional AI enrichment, official upstream allowlisting, mirror diff/explain reporting, offline workspace lifecycle smoke coverage, adapter-driven native-install/runtime preflight, `.gitignore` glob/negation handling, and shared CLI option parsing that rejects flag tokens as missing values.
+Follow-up hardening after the merge addressed high-risk audit findings: scoped public package identity as `@ar27111994/agent-harness`, deterministic package allowlisting and packed-artifact smoke checks, package `main`/`exports`/`types` entry points, mutable state-root support through `--state-root` and `AGENT_HARNESS_STATE_ROOT`, safe mirror path resolution and file-manifest verification for multi-file artifacts, bounded mirror evidence file reads under approved local roots, guarded external response fetching with origin allowlists and byte limits, static public-provider allowlisting for optional AI enrichment, GitHub/repo content mirroring during mirror acquisition instead of wire time, validated PyPI metadata normalization, lazy VS Code settings path resolution after `.env` loading, resettable process-local GitHub state, explicit recommendation execution in the workspace and full rebuild pipelines, registry-driven recommendation-host enumeration, adapter-derived install bundle lock discovery, safe preview-by-default wire mode, non-destructive VS Code/OpenCode managed sections, quarantine review commands with audit logging, optional AI enrichment, official upstream allowlisting, mirror diff/explain reporting, offline workspace lifecycle smoke coverage, adapter-driven native-install/runtime preflight, `.gitignore` glob/negation handling, recommendation and mirror module decomposition, type-aware ESLint hardening, and shared CLI option parsing that rejects flag tokens as missing values.
 
 Validation for this release includes typecheck, lint, format check, build, unit/link lifecycle tests, dotenv/parser tests, state-root tests, VS Code settings tests, security hardening tests, scan benchmark budget checks, detection quality reporting, detector-to-policy coverage reporting, isolated CLI smoke checks, offline workspace lifecycle smoke checks, packed-artifact smoke checks, recommendation fixture evaluation, and a Windows/macOS/Linux CI matrix plus a provenance-ready release workflow.
