+++ Roadmap.md
@@ -20,7 +20,7 @@ Scope of the original assessment:
 
 The `release/v1.0.0` execution wave closes the highest-priority roadmap gaps for portability, generic discovery, source utilization, measurable detection quality, host adapter extensibility, and native project-local wire-in. In particular, VS Code/Copilot and OpenCode have native wire-in flows, while Cursor, Zed, Claude Code, and Pi now use the same host adapter model to write project-local native host files, host-specific wire plans, and independently tuned recommendation policies.
 
-The post-merge hardening wave adds scoped package release identity, deterministic package allowlisting, package entry points, mutable state-root support outside the package install directory, safe mirror writes with file-manifest verification, bounded mirror evidence file reads, guarded network reads, static public-provider allowlisting for optional AI enrichment, GitHub/repo mirror-time content acquisition, validated PyPI metadata, lazy `.env`-aware VS Code settings path resolution, explicit recommendation execution in workspace and full rebuild runs, registry-driven recommendation-host enumeration, adapter-derived install bundle discovery, preview-by-default wire mode, non-destructive VS Code/OpenCode managed sections, adapter-owned runtime checks, VS Code native extension install/verify/remove with bounded native commands, asset prerequisite guidance, optional AI enrichment, official upstream allowlisting, mirror diff/explain reporting, isolated CLI, offline workspace, and packed-artifact smoke checks, operational docs/registry/marketplace harvesters, broad deterministic file-family and technology signature packs, `.gitignore` glob/negation handling, focused install-domain modules, domain-specific manifest validators, localized domain type modules, discovery demand/source/reporting seams, and package/reference/local/GitHub/official-index catalog harvester seams. Broader native installers beyond VS Code and Cursor, richer interactive quarantine review UIs, deeper provider-specific OAuth automation, stricter lint policy, remaining oversized-file decomposition, and future package/workspace extraction remain tracked follow-up work.
+The post-merge hardening wave adds scoped package release identity, deterministic package allowlisting, package entry points and declaration output, mutable state-root support outside the package install directory, safe mirror writes with file-manifest verification, bounded mirror evidence file reads, guarded network reads, static public-provider allowlisting for optional AI enrichment, GitHub/repo mirror-time content acquisition, validated PyPI metadata, lazy `.env`-aware VS Code settings path resolution, explicit recommendation execution in workspace and full rebuild runs, registry-driven recommendation-host enumeration, adapter-derived install bundle discovery, preview-by-default wire mode, non-destructive VS Code/OpenCode managed sections, adapter-owned runtime checks, VS Code native extension install/verify/remove with bounded native commands, asset prerequisite guidance, optional AI enrichment, official upstream allowlisting, mirror diff/explain reporting, isolated CLI, offline workspace, and packed-artifact smoke checks, operational docs/registry/marketplace harvesters, broad deterministic file-family and technology signature packs, `.gitignore` glob/negation handling, focused install-domain modules, recommendation and mirror module seams, domain-specific manifest validators, localized domain type modules, discovery demand/source/reporting seams, package/reference/local/GitHub/official-index catalog harvester seams, and type-aware ESLint guardrails. Broader native installers beyond VS Code and Cursor, richer interactive quarantine review UIs, deeper provider-specific OAuth automation, and future package/workspace extraction remain tracked follow-up work.
 
 The gap matrices below are retained for traceability. Their evidence cells describe the original 2026-04-19 baseline, while the verdict table summarizes the current status.
