+++ eslint.config.mjs
@@ -4,6 +4,32 @@ import { defineConfig } from "eslint/config";
 import globals from "globals";
 import tseslint from "typescript-eslint";
 
+const cliOutputFiles = [
+  "src/activate.ts",
+  "src/cli.ts",
+  "src/discover.ts",
+  "src/domains/discovery/ai-enrichment.ts",
+  "src/domains/discovery/catalog-inspection.ts",
+  "src/domains/discovery/source-index.ts",
+  "src/install.ts",
+  "src/install/*.ts",
+  "src/mirror.ts",
+  "src/mirror/*.ts",
+  "src/quarantine.ts",
+  "src/rebuild.ts",
+  "src/recommend.ts",
+  "src/recommend/commands.ts",
+  "src/setup.ts",
+  "src/wire.ts",
+  "src/workspace.ts",
+  "src/tests/cli-smoke.ts",
+  "src/tests/detection-quality.ts",
+  "src/tests/pack-smoke.ts",
+  "src/tests/policy-coverage.ts",
+  "src/tests/scan-benchmark.ts",
+  "src/tests/workspace-smoke.ts",
+];
+
 export default defineConfig(
   {
     ignores: [
@@ -36,6 +62,12 @@ export default defineConfig(
   ...tseslint.configs.recommended,
   {
     files: ["src/**/*.ts"],
+    languageOptions: {
+      parserOptions: {
+        projectService: true,
+        tsconfigRootDir: import.meta.dirname,
+      },
+    },
     rules: {
       "@typescript-eslint/consistent-type-imports": [
         "error",
@@ -44,6 +76,16 @@ export default defineConfig(
           fixStyle: "separate-type-imports",
         },
       ],
+      "@typescript-eslint/no-floating-promises": "error",
+      "@typescript-eslint/no-unsafe-assignment": "error",
+      "@typescript-eslint/no-unsafe-member-access": "error",
+      "no-console": ["error", { allow: ["warn", "error"] }],
+    },
+  },
+  {
+    files: cliOutputFiles,
+    rules: {
+      "no-console": ["error", { allow: ["log", "warn", "error"] }],
     },
   },
   prettierConfig,
