+++ tsconfig.json
@@ -8,7 +8,7 @@
     "outDir": "dist",
     "types": ["node"],
     "sourceMap": true,
-    "declaration": false,
+    "declaration": true,
     "noEmitOnError": true,
     "skipLibCheck": true
   },
